backend/requirements.txt

fastapi
uvicorn
python-docx
pdfminer.six
openai
python-multipart
pydantic
aiofiles
